+++ .vscode/settings.json
@@ -3,6 +3,7 @@
     "files.associations": {
         "mlx42.h": "c",
         "cub3d.h": "c",
-        "ostream": "c"
+        "ostream": "c",
+        "complex": "c"
     }
 }
