+++ maps/map.cub
@@ -10,11 +10,11 @@ F 128,128,128
 111111111111111
 100000000000001
 100000111111001
-100100122221001
+1001001    1001
 100100111111001
 100100000000001
-100000000000001
+1000S0000000001
 100000000001001
-10000100W001001
+100001001001001
 100000000000001
 111111111111111
