+++ maps/test.cub
@@ -9,11 +9,11 @@ F 128,128,128
 
 
 
-    11111111111111111111111111111111
+11111111111111111111111111111111111
 111110000000000000000000000000000011
-    11100000000011111111111111000001
-11111111111111001             111111
-    111000000S001             100001
+111000000000111111111111111000011111
+11111111111111111          11111111111
+111000000S001111111           100001
 111111110000000111111111111111000011
-    11100000000000000000000000011111
-    11111111111111111111111111111111
+11100000000000000000000000011111111
+111111111111111111111111111111
