+++ tools/ft_isalpha.c
@@ -1,0 +1,10 @@
+
+#include "../cub3d.h"
+
+int	ft_isalpha(int c)
+{
+	if ((c >= 65 && c <= 90) || (c >= 97 && c <= 122))
+		return (1);
+	else
+		return (0);
+}
